python/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpha_pipeline"
version = "0.1.0"
description = "Pipeline evolutivo híbrido para diseño de microservicios"
authors = [
    { name = "Daniel Narváez", email = "JoseDaniel.NarvaezF@alumnos.uai.edu.ar" }
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "openai>=0.27.0",
    "nltk>=3.8",
    "networkx>=3.1",
    "numpy>=1.25",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2",
    "flake8>=5.0",
    "mypy>=1.0"
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["alpha_pipeline*"]
